requirements.txt
langchain==0.2.6
langchain-community==0.2.6
sentence-transformers==2.6.1
scikit-learn==0.22.2.post1
coclust==0.2.1
rouge-score==0.1.2
scipy==1.13.0
numpy==1.23.5
tqdm==4.66.4
torch==2.2.2